Attack and Ship Boards are now displayed
Increased game resolution from 1280x720 to 1600x900 for additional space for boards. Also changed the X_OFFSET variable to change the position of the boards.

During a player's turn, the left grid will be used for attacking and the right board displays the current player's ships (green circle indicates placements for player's ships) and enemy misses/hits.

Added an the argument if_ships_visible to the draw_board function to indicate if the board should show ships or not. This is used so the current player does not see the enemy ships in the player_turn function.